Starting to check all system calls

+++ src/events/interface.h
@@ -1,6 +1,9 @@
 #ifndef INTERFACE_H
 #define INTERFACE_H
 
+#include "defines.h"
+#include "error.h"
+
 /**
  * 
  * send buf with brigthness on i2c file
@@ -16,6 +19,6 @@ void i2c_send(const devices_info *dev_info);
  * @param int events val
  *
  **/
-void emit(const devices_info *dev_info, int type, int code, int val);
+void emit(const devices_info *dev_info, unsigned short type, unsigned short code, int val);
 
 #endif // INTERFACE_H
